pyproject.toml
[tool.poetry]
name = "python-duckdb-etl"
version = "0.1.0"
description = ""
authors = ["Robinson Lovatto <robinson.lovatto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.35.0"
gdown = "^5.2.0"
duckdb = "^1.0.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
